+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - selector handling for `RETRY_CSS_SELECTORS` in `_handle_blocked_request` in `BeautifulSoupCrawler`
+- selector handling in `enqueue_links` in `BeautifulSoupCrawler`
 
 ## [0.0.6](../../releases/tag/v0.0.6) - 2024-06-25